Switch to using mapstructure for response parsing
Changing to `mapstructure` parsing rather than the default `json` parsing gives us leeway
in cases where the OpenAPI spec is out of date or inaccurate.

The approach we're taking here is to unmarshal the json response into a `map[string]any`
and afterwards use `mapstructure` to convert the map into the desired struct.
`mapstructure` does not stop unmarshalling when it encounters an error, instead all
errors are joined together and returned at the end.

Prior to this change we would not be able to get any data out of the response if there were
even minor differences from the OpenAPI spec.